+++ src/gromacs/nbnxm/atomdata.h
@@ -386,7 +386,7 @@ void nbnxn_atomdata_set(nbnxn_atomdata_t*            nbat,
                         const Nbnxm::GridSet&        gridSet,
                         gmx::ArrayRef<const int>     atomTypes,
                         gmx::ArrayRef<const real>    atomCharges,
-                        gmx::ArrayRef<const int64_t> atomInfo);
+                        gmx::ArrayRef<const int32_t> atomInfo);
 
 //! Copy the shift vectors to nbat
 void nbnxn_atomdata_copy_shiftvec(bool dynamic_box, gmx::ArrayRef<gmx::RVec> shift_vec, nbnxn_atomdata_t* nbat);
